pyproject.toml
[tool.poetry]
name = "analsis-recipes"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
jupyter = "^1.1.1"
itables = "^2.4.2"
polars = "^1.30.0"
cas-visualizer = "^0.1.1"
dakoda-core = {path = "../dakoda-core/"}
xsdata = "22.12"
textstat = "^0.7.3"
seaborn = "^0.13.2"
spacy = "3.8.7"
CairoSVG = "2.8.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "."
]